Resolve #9 by restructuring source and removing dependency on Babel
