+++ .gitignore
@@ -1,5 +1,4 @@
 npm-debug.log
 .russrc
 russ.tasks/
-test/
 node_modules/
