+++ .travis.yml
@@ -1,10 +1,8 @@
 language: node_js
 node_js:
-  - "0.10"
+  - "6"
 sudo: false
 install:
   - npm install
-before_script:
-  - make build
 script:
   - make run-tests
