+++ Makefile
@@ -1,7 +1,6 @@
 PHONY: help
 
 MODULES = ./node_modules/.bin
-BABEL = $(MODULES)/babel
 MOCHA = $(MODULES)/mocha
 ESLINT = $(MODULES)/eslint
 
@@ -14,19 +13,10 @@ help:
 	@grep -E '^[a-zA-Z\._-]+:.*?## .*$$' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-20s\033[0m %s\n", $$1, $$2}'
 
 setup: ## Set up project
-	npm install
-
-build: ## Build bolt
-	$(BABEL) $(SRC) -d $(DEST)
-
-watch: build ## Watcher for babel
-	$(BABEL) $(SRC) --watch -d $(DEST)
+	yarn
 
 lint: ## Lint src using es2015
 	$(ESLINT) $(SRC)
 
-dev:## Develop bolt
-	make watch
-
-run-tests: ## Run bolt tests
+run-tests: ## Run tests
 	$(MOCHA) $(TESTOPTS)
