+++ lib/core/instance.js
@@ -1,13 +1,6 @@
-'use strict';
-
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-require('babel-polyfill');
-var winston = require('winston'),
-    fs = require('fs'),
-    RussTask = require('./task');
+const winston  = require('winston'),
+  fs           = require('fs'),
+  RussTask     = require('./task');
 
 /**
   * @class RussInstance
@@ -16,15 +9,12 @@ var winston = require('winston'),
   * @param env {string} - define runtime environment
   * @returns {Object} - russInstance
 */
-
-var RussInstance = function () {
-  function RussInstance(env) {
-    _classCallCheck(this, RussInstance);
-
-    var tasks = void 0;
-    var config = void 0;
+class RussInstance {
+  constructor(env) {
+    let tasks;
+    let config;
     try {
-      config = require(process.cwd() + '/.russrc');
+      config = require(`${process.cwd()}/.russrc`);
       tasks = fs.readdirSync('russ.tasks');
     } catch (err) {
       throw Error('Missing russ files...');
@@ -35,179 +25,137 @@ var RussInstance = function () {
     return this;
   }
 
-  _createClass(RussInstance, [{
-    key: 'run',
-    value: function run(taskPool) {
-      var _this = this;
-
-      var tasks = taskPool[Symbol.iterator]();
-      var exec = function exec(name, resolve, reject) {
-        var cb = function cb() {
-          winston.profile(name);
-          if (tasks.next) {
-            var nextTask = tasks.next();
-            if (!nextTask.done) exec(nextTask.value, resolve, reject);else resolve();
-          }
-        };
-        var errCb = function errCb(err) {
-          if (err) winston.error('Error: ' + err);
-          reject(err);
-        };
-        try {
-          var taskObj = _this.tasks[name];
-          if (taskObj.concurrent) {
-            var _tasks = taskObj.concurrent.map(_this.runTask.bind(_this));
-            Promise.all(_tasks).then(cb).catch(errCb);
-          } else {
-            var task = void 0;
-            try {
-              task = new RussTask(_this, _this.tasks[name]);
-            } catch (err) {
-              winston.error(err.toString());
-              reject(err.toString());
-            }
-            task.run(_this.env).then(cb).catch(errCb);
-          }
-        } catch (err) {
-          reject(err);
+  run(taskPool) {
+    const tasks = taskPool[Symbol.iterator]();
+    const exec = (name, resolve, reject) => {
+      const cb = () => {
+        winston.profile(name);
+        if (tasks.next) {
+          const nextTask = tasks.next();
+          if (!nextTask.done)
+            exec(nextTask.value, resolve, reject);
+          else
+            resolve();
         }
       };
-      return new Promise(function (resolve, reject) {
-        exec(tasks.next().value, resolve, reject);
-      });
-    }
-  }, {
-    key: 'getPool',
-    value: function getPool(name) {
-      var _this2 = this;
-
-      var pool = [];
-      var pushToPool = function pushToPool(name, parent) {
-        if (pool.indexOf(name) !== -1) return;
-        var task = _this2.tasks[name];
-        if (!task) throw Error('Task ' + name + ' is not defined...');
-        var clean = function clean(a) {
-          return a && pool.indexOf(a) === -1;
-        };
-        var tasks = void 0;
-        if (task.sequence) tasks = task.sequence.filter(clean);else if (!task.concurrent) tasks = [task.pre, task.post].filter(clean);
-
-        if (parent) {
-          var parentTask = _this2.tasks[parent];
-          var pIdx = pool.indexOf(parent);
-          var idx = name === parentTask.pre ? pIdx : pIdx + 1;
-          pool.splice(idx, 0, name);
-        } else if (!task.sequence) pool.push(name);
-
-        var newParent = task.sequence ? undefined : name;
-        if (tasks && tasks.length > 0) {
-          var _iteratorNormalCompletion = true;
-          var _didIteratorError = false;
-          var _iteratorError = undefined;
-
-          try {
-            for (var _iterator = tasks[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {
-              var _t = _step.value;
-              pushToPool(_t, newParent);
-            }
-          } catch (err) {
-            _didIteratorError = true;
-            _iteratorError = err;
-          } finally {
-            try {
-              if (!_iteratorNormalCompletion && _iterator.return) {
-                _iterator.return();
-              }
-            } finally {
-              if (_didIteratorError) {
-                throw _iteratorError;
-              }
-            }
-          }
-        }
+      const errCb = (err) => {
+        if (err) winston.error(`Error: ${err}`);
+        reject(err);
       };
-      pushToPool(name);
-      return pool;
-    }
-  }, {
-    key: 'runTask',
-    value: function runTask(name) {
-      var _this3 = this;
-
-      var task = this.tasks[name];
-      if (!task) throw Error('No such task...');
-      var taskPool = this.getPool(name);
-      return new Promise(function (resolve, reject) {
-        if (task.sequence || task.concurrent) {
-          winston.info('Running ' + task.name);
-          winston.profile(task.name);
-        }
-        _this3.run(taskPool).then(function () {
-          if (task.sequence || task.concurrent) winston.profile(task.name);
-          resolve();
-        }).catch(function (err) {
-          reject(err);
-        });
-      });
-    }
-  }, {
-    key: 'info',
-    value: function info() {
-      var header = 'Available task to run:\n';
-      var taskList = '\n';
-      if (Object.keys(this.tasks).length > 0) for (var task in this.tasks) {
-        taskList += '     ' + task.green + ': ' + this.tasks[task].doc.cyan + '\n';
-      }winston.info('' + header + taskList);
-    }
-
-    /**
-      * populates instance tasks
-      * @param files {Array} array of task objects to use when registering
-    */
-
-  }, {
-    key: 'register',
-    value: function register(files) {
-      var _this4 = this;
-
-      var registerTask = function registerTask(opts) {
-        var ERR_MSG = 'Task ' + opts.name + ' missing properties...';
-        var hasFunc = opts.func && typeof opts.func === 'function';
-        var isDel = opts.concurrent || opts.sequence;
-        if (opts.name && opts.doc && (hasFunc || isDel)) _this4.tasks[opts.name] = opts;else throw new Error(ERR_MSG);
-      };
-      if (files.length === 0) throw new Error('No tasks defined in russ.tasks');
-      this.tasks = {};
-      var _iteratorNormalCompletion2 = true;
-      var _didIteratorError2 = false;
-      var _iteratorError2 = undefined;
-
       try {
-        for (var _iterator2 = files[Symbol.iterator](), _step2; !(_iteratorNormalCompletion2 = (_step2 = _iterator2.next()).done); _iteratorNormalCompletion2 = true) {
-          var file = _step2.value;
-
-          var taskOpts = require(process.cwd() + '/russ.tasks/' + file);
-          if (this.tasks[taskOpts.name]) throw new Error('Task ' + taskOpts.name + ' already defined...');
-          if (Array.isArray(taskOpts)) taskOpts.map(registerTask);else registerTask(taskOpts);
+        const taskObj = this.tasks[name];
+        if (taskObj.concurrent) {
+          const tasks = taskObj.concurrent.map(this.runTask.bind(this));
+          Promise.all(tasks)
+            .then(cb)
+            .catch(errCb);
+        } else {
+          let task;
+          try {
+            task = new RussTask(this, this.tasks[name]);
+          } catch (err) {
+            winston.error(err.toString());
+            reject(err.toString());
+          }
+          task.run(this.env)
+            .then(cb)
+            .catch(errCb);
         }
       } catch (err) {
-        _didIteratorError2 = true;
-        _iteratorError2 = err;
-      } finally {
-        try {
-          if (!_iteratorNormalCompletion2 && _iterator2.return) {
-            _iterator2.return();
-          }
-        } finally {
-          if (_didIteratorError2) {
-            throw _iteratorError2;
-          }
-        }
+        reject(err);
       }
-    }
-  }]);
+    };
+    return new Promise((resolve, reject) => {
+      exec(tasks.next().value, resolve, reject);
+    });
+  }
 
-  return RussInstance;
-}();
+  getPool(name) {
+    const pool = [];
+    const pushToPool = (name, parent) => {
+      if (pool.indexOf(name) !== -1) return;
+      const task = this.tasks[name];
+      if (!task) throw Error(`Task ${name} is not defined...`);
+      const clean = (a) => {
+        return a && (pool.indexOf(a) === -1);
+      };
+      let tasks;
+      if (task.sequence)
+        tasks = task.sequence.filter(clean);
+      else if (!task.concurrent)
+        tasks = [task.pre, task.post].filter(clean);
+
+      if (parent) {
+        const parentTask = this.tasks[parent];
+        const pIdx = pool.indexOf(parent);
+        const idx = (name === parentTask.pre) ? pIdx : (pIdx + 1);
+        pool.splice(idx, 0, name);
+      } else if (!task.sequence) pool.push(name);
+
+      const newParent = (task.sequence) ? undefined : name;
+      if (tasks && tasks.length > 0)
+        for (const t of tasks) pushToPool(t, newParent);
+    };
+    pushToPool(name);
+    return pool;
+  }
+
+  runTask(name) {
+    const task = this.tasks[name];
+    if (!task) throw Error('No such task...');
+    const taskPool = this.getPool(name);
+    return new Promise((resolve, reject) => {
+      if (task.sequence || task.concurrent) {
+        winston.info(`Running ${task.name}`);
+        winston.profile(task.name);
+      }
+      this.run(taskPool)
+        .then(() => {
+          if (task.sequence || task.concurrent)
+            winston.profile(task.name);
+          resolve();
+        })
+        .catch((err) => {
+          reject(err);
+        });
+    });
+  }
+
+  info() {
+    const header = 'Available task to run:\n';
+    let taskList = '\n';
+    if (Object.keys(this.tasks).length > 0)
+      for (const task in this.tasks)
+        taskList += `     ${task.green}: ${this.tasks[task].doc.cyan}\n`;
+    winston.info(`${header}${taskList}`);
+  }
+
+  /**
+    * populates instance tasks
+    * @param files {Array} array of task objects to use when registering
+  */
+  register(files) {
+    const registerTask = (opts) => {
+      const ERR_MSG = `Task ${opts.name} missing properties...`;
+      const hasFunc = opts.func && typeof opts.func === 'function';
+      const isDel   = opts.concurrent || opts.sequence;
+      if (opts.name && opts.doc && (hasFunc || isDel))
+        this.tasks[opts.name] = opts;
+      else
+        throw new Error(ERR_MSG);
+    };
+    if (files.length === 0) throw new Error('No tasks defined in russ.tasks');
+    this.tasks = {};
+    for (const file of files) {
+      const taskOpts = require(`${process.cwd()}/russ.tasks/${file}`);
+      if (this.tasks[taskOpts.name])
+        throw new Error(`Task ${taskOpts.name} already defined...`);
+      if (Array.isArray(taskOpts))
+        taskOpts.map(registerTask);
+      else
+        registerTask(taskOpts);
+    }
+  }
+}
 
 module.exports = RussInstance;
