+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "russ",
   "description": "a node script runner",
-  "version": "0.9.3",
+  "version": "0.10.0",
   "main": "index.js",
   "licenses": [
     {
@@ -18,14 +18,10 @@
   "dependencies": {
     "colors": "^1.1.2",
     "commander": "^2.9.0",
-    "winston": "^2.2.0",
-    "babel-polyfill": "^6.7.4"
+    "winston": "^2.2.0"
   },
   "keywords": [],
   "devDependencies": {
-    "babel-cli": "^6.7.5",
-    "babel-eslint": "^6.0.2",
-    "babel-preset-es2015": "^6.6.0",
     "chai": "^3.5.0",
     "eslint": "^2.8.0",
     "mkdirp": "^0.5.1",
